support LtFileV1 and LtFileV2
This adds support for both v1 and v2 LT file formats.
